# static/swagger-docs.json
{"openapi":"3.0.0","paths":{"/api/system/menu":{"post":{"operationId":"MenuController_create","summary":"新建菜单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMenuDto"}}}},"responses":{"201":{"description":"新建菜单内容","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuEntity"}}}}},"tags":["系统管理-菜单"]}},"/api/system/menu/list":{"get":{"operationId":"MenuController_findList","summary":"菜单列表","parameters":[],"responses":{"200":{"description":"菜单内容","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuEntity"}}}}}},"tags":["系统管理-菜单"]}},"/api/system/menu/tree-list":{"get":{"operationId":"MenuController_findTreeList","summary":"菜单树列表","parameters":[],"responses":{"200":{"description":"菜单内容 tree格式","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuTreeListVo"}}}}}},"tags":["系统管理-菜单"]}},"/api/system/menu/{id}":{"get":{"operationId":"MenuController_findOne","summary":"根据id查菜单详情（不含子级)）","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"菜单详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuEntity"}}}}},"tags":["系统管理-菜单"]},"patch":{"operationId":"MenuController_update","summary":"更新菜单详情","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMenuDto"}}}},"responses":{"200":{"description":"菜单详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuEntity"}}}}},"tags":["系统管理-菜单"]},"delete":{"operationId":"MenuController_remove","summary":"删除菜单","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["系统管理-菜单"]}},"/api/system/account":{"post":{"operationId":"AccountController_create","summary":"创建系统用户","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountDto"}}}},"responses":{"201":{"description":"创建成功的系统用户信息","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountVo"}}}}},"tags":["系统管理-用户"]},"get":{"operationId":"AccountController_findAll","summary":"查看系统用户。分页","parameters":[{"name":"pageNo","required":true,"in":"query","description":"页码","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","description":"每页大小","schema":{"type":"number"}}],"responses":{"200":{"description":"分页查询信息","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountListVo"}}}}},"tags":["系统管理-用户"]}},"/api/system/account/permissions":{"get":{"operationId":"AccountController_getPermissions","summary":"获取用户权限","parameters":[],"responses":{"200":{"description":"权限列表","content":{"application/json":{"schema":{"type":"array"}}}}},"tags":["系统管理-用户"]}},"/api/system/account/{id}":{"get":{"operationId":"AccountController_findOne","summary":"查询用户详细","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"用户详细信息","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountVo"}}}}},"tags":["系统管理-用户"]},"patch":{"operationId":"AccountController_update","summary":"更改用户信息","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAccountDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["系统管理-用户"]},"delete":{"operationId":"AccountController_remove","summary":"删除用户","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["系统管理-用户"]}},"/api/system/account/{id}/set-password":{"post":{"operationId":"AccountController_setPassword","summary":"设置用户密码","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetPasswordAccountDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}},"201":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["系统管理-用户"]}},"/api/system/account/{id}/reset-password":{"post":{"operationId":"AccountController_resetPassword","summary":"重置用户密码","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordAccountDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}},"201":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["系统管理-用户"]}},"/api/system/role":{"post":{"operationId":"RoleController_create","summary":"新增角色","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":"角色详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["系统管理-角色"]},"get":{"operationId":"RoleController_findAll","summary":"查询角色，分页","parameters":[{"name":"pageNo","required":true,"in":"query","description":"页码","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","description":"每页大小","schema":{"type":"number"}}],"responses":{"200":{"description":"分页查询信息","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleListVo"}}}}},"tags":["系统管理-角色"]}},"/api/system/role/simple-list":{"get":{"operationId":"RoleController_findAllSimple","summary":"查询角色,枚举","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleSimpleListVo"}}}}}},"tags":["系统管理-角色"]}},"/api/system/role/{id}":{"get":{"operationId":"RoleController_findOne","summary":"查询角色详情","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["系统管理-角色"]},"patch":{"operationId":"RoleController_update","summary":"更新角色信息","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["系统管理-角色"]},"delete":{"operationId":"RoleController_remove","summary":"删除角色","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["系统管理-角色"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","summary":"登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"登录成功,返回token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginVo"}}}}},"tags":["auth"]}},"/api/auth/encryption":{"post":{"operationId":"AuthController_encryption","summary":"加密字符串","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EncryptionDto"}}}},"responses":{"200":{"description":"返回加密后的字符串","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["auth"]}},"/api/auth/refresh-token":{"post":{"operationId":"AuthController_refreshToken","summary":"刷新token","parameters":[],"responses":{"200":{"description":"刷新成功，返回token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginVo"}}}}},"tags":["auth"]}},"/api/log/login-log":{"post":{"operationId":"LoginLogController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLoginLogDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginLogEntity"}}}}},"tags":["日志管理-登录日志"]},"get":{"operationId":"LoginLogController_findAll","summary":"查看登录日志。分页","parameters":[{"name":"pageNo","required":true,"in":"query","description":"页码","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","description":"每页大小","schema":{"type":"number"}},{"name":"startDate","required":false,"in":"query","description":"开始时间。格式 Date","schema":{"format":"date-time","type":"string"}},{"name":"endDate","required":false,"in":"query","description":"结束时间。格式 Date","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":"分页查询信息","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginLogListVo"}}}}},"tags":["日志管理-登录日志"]}},"/api/other/swagger-docs":{"get":{"operationId":"OtherController_getFile","parameters":[],"responses":{"200":{"description":""}}}},"/api/other/get-server-info":{"get":{"operationId":"OtherController_getServerInfo","summary":"获取服务器信息","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInfoVo"}}}}}}}},"info":{"title":"mu后台管理系统 api 文档","description":"api接口文档","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateMenuDto":{"type":"object","properties":{"type":{"type":"number","description":"类型。1:目录,2:菜单,3:按钮","default":1,"enum":[1,2]},"show":{"type":"number","description":"是否显示。1:显示,0:不显示，默认1","default":1,"enum":[1,0]},"name":{"type":"string","description":"名称","maxLength":32},"parentId":{"type":"number","description":"父级id"},"sort":{"type":"number","description":"排序。最大值 9999","maximum":9999},"accessCode":{"type":"string","description":"权限码","maxLength":64}},"required":["type","name","accessCode"]},"MenuEntity":{"type":"object","properties":{"type":{"type":"number","description":"类型。1:目录,2:菜单,3:按钮","enum":[1,2]},"show":{"type":"number","description":"是否显示。1:显示,0:不显示，默认1","default":1,"enum":[1,0]},"parentId":{"type":"number","description":"父级id"},"name":{"type":"string","description":"名称"},"accessCode":{"type":"string","description":"权限码"},"sort":{"type":"number","description":"排序"},"id":{"type":"number","description":"唯一标识"},"createdAt":{"format":"date-time","type":"string","description":"创建时间"},"updatedAt":{"format":"date-time","type":"string","description":"更新时间"}},"required":["type","show","name","accessCode","id","createdAt","updatedAt"]},"MenuTreeListVo":{"type":"object","properties":{"type":{"type":"number","description":"类型。1:目录,2:菜单,3:按钮","enum":[1,2]},"show":{"type":"number","description":"是否显示。1:显示,0:不显示，默认1","default":1,"enum":[1,0]},"children":{"description":"子级","type":"array","items":{"$ref":"#/components/schemas/MenuEntity"}},"parentId":{"type":"number","description":"父级id"},"name":{"type":"string","description":"名称"},"accessCode":{"type":"string","description":"权限码"},"sort":{"type":"number","description":"排序"},"id":{"type":"number","description":"唯一标识"},"createdAt":{"format":"date-time","type":"string","description":"创建时间"},"updatedAt":{"format":"date-time","type":"string","description":"更新时间"}},"required":["type","show","name","accessCode","id","createdAt","updatedAt"]},"UpdateMenuDto":{"type":"object","properties":{"type":{"type":"number","description":"类型。1:目录,2:菜单,3:按钮","default":1,"enum":[1,2]},"show":{"type":"number","description":"是否显示。1:显示,0:不显示，默认1","default":1,"enum":[1,0]},"name":{"type":"string","description":"名称","maxLength":32},"parentId":{"type":"number","description":"父级id"},"sort":{"type":"number","description":"排序。最大值 9999","maximum":9999},"accessCode":{"type":"string","description":"权限码","maxLength":64}}},"CreateAccountDto":{"type":"object","properties":{"status":{"type":"number","description":"状态。1:正常，0: 停用'","enum":[1,0],"default":1},"roleIds":{"description":"角色ids","type":"array","items":{"type":"number"}},"userName":{"type":"string","description":"用户名","maxLength":32},"name":{"type":"string","description":"姓名","maxLength":32},"nickname":{"type":"string","description":"昵称","maxLength":32},"email":{"type":"string","description":"邮箱","maxLength":64},"mobile":{"type":"string","description":"手机号","maxLength":48},"remark":{"type":"string","description":"备注","maxLength":64}},"required":["status","roleIds","userName"]},"PickTypeClass":{"type":"object","properties":{"name":{"type":"string","description":"角色名称"},"id":{"type":"number","description":"唯一标识"}},"required":["name","id"]},"AccountVo":{"type":"object","properties":{"status":{"type":"number","description":"状态。1:正常，0: 停用","enum":[1,0],"default":1},"roleList":{"description":"角色集合","type":"array","items":{"$ref":"#/components/schemas/PickTypeClass"}},"id":{"type":"number","description":"唯一标识"},"createdAt":{"format":"date-time","type":"string","description":"创建时间"},"updatedAt":{"format":"date-time","type":"string","description":"更新时间"},"userName":{"type":"string","description":"用户名,唯一"},"name":{"type":"string","description":"姓名"},"nickname":{"type":"string","description":"昵称"},"email":{"type":"string","description":"邮箱"},"mobile":{"type":"string","description":"手机号"},"remark":{"type":"string","description":"备注"}},"required":["status","roleList","id","createdAt","updatedAt","userName"]},"AccountListVo":{"type":"object","properties":{"total":{"type":"number","description":"列表总条数"},"pageSize":{"type":"number","description":"分页大小"},"pageNo":{"type":"number","description":"页码"},"list":{"description":"列表数据。","type":"array","items":{"$ref":"#/components/schemas/AccountVo"}}},"required":["total","pageSize","pageNo","list"]},"UpdateAccountDto":{"type":"object","properties":{"status":{"type":"number","description":"状态。1:正常，0: 停用'","enum":[1,0],"default":1},"roleIds":{"description":"角色ids","type":"array","items":{"type":"number"}},"userName":{"type":"string","description":"用户名","maxLength":32},"name":{"type":"string","description":"姓名","maxLength":32},"nickname":{"type":"string","description":"昵称","maxLength":32},"email":{"type":"string","description":"邮箱","maxLength":64},"mobile":{"type":"string","description":"手机号","maxLength":48},"remark":{"type":"string","description":"备注","maxLength":64}}},"SetPasswordAccountDto":{"type":"object","properties":{"oldPassword":{"type":"string","maxLength":128,"description":"用户旧密码。需经加密"},"newPassword":{"type":"string","maxLength":128,"description":"用户新密码。需经加密"}},"required":["oldPassword","newPassword"]},"ResetPasswordAccountDto":{"type":"object","properties":{"password":{"type":"string","maxLength":128,"description":"用户新密码。需经加密"}}},"CreateRoleDto":{"type":"object","properties":{"name":{"type":"string","description":"角色名称","maxLength":32},"remark":{"type":"string","description":"备注","maxLength":64},"status":{"type":"object","description":"状态。1:正常，0: 停用"},"menuIds":{"type":"string","description":"菜单权限id。多个用,分隔"}},"required":["name","status"]},"RoleEntity":{"type":"object","properties":{"status":{"type":"number","description":"状态。1:正常，0: 停用","enum":[1,0],"default":1},"name":{"type":"string","description":"角色名称"},"remark":{"type":"string","description":"备注"},"menuIds":{"type":"string","description":"菜单权限id"},"id":{"type":"number","description":"唯一标识"},"createdAt":{"format":"date-time","type":"string","description":"创建时间"},"updatedAt":{"format":"date-time","type":"string","description":"更新时间"}},"required":["status","name","id","createdAt","updatedAt"]},"RoleListVo":{"type":"object","properties":{"total":{"type":"number","description":"列表总条数"},"pageSize":{"type":"number","description":"分页大小"},"pageNo":{"type":"number","description":"页码"},"list":{"description":"列表数据。","type":"array","items":{"$ref":"#/components/schemas/RoleEntity"}}},"required":["total","pageSize","pageNo","list"]},"RoleSimpleListVo":{"type":"object","properties":{"value":{"type":"number","description":"role id"},"label":{"type":"string","description":"role name"}},"required":["value","label"]},"UpdateRoleDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{"userName":{"type":"string","description":"账号"},"password":{"type":"string","description":"密码。需经过接口 auth/encryption 加密过的字符串"}},"required":["userName","password"]},"LoginVo":{"type":"object","properties":{"accessToken":{"type":"string","description":"用户认证token"}},"required":["accessToken"]},"EncryptionDto":{"type":"object","properties":{"str":{"type":"string","description":"需加密字符串"}},"required":["str"]},"CreateLoginLogDto":{"type":"object","properties":{"status":{"type":"number","description":"状态。1:成功，0: 失败","enum":[1,0],"default":1},"userName":{"type":"string","description":"登录账号"},"ip":{"type":"string","description":"ip"},"equipment":{"type":"string","description":"登录设备"},"info":{"type":"string","description":"登录信息"},"userAgent":{"type":"string"}},"required":["status","userName","ip"]},"LoginLogEntity":{"type":"object","properties":{"status":{"type":"number","description":"状态。1:成功，0: 失败","enum":[1,0],"default":1},"userName":{"type":"string","description":"登录账号"},"ip":{"type":"string","description":"登陆地址"},"equipment":{"type":"string","description":"登录设备"},"info":{"type":"string","description":"登录信息"},"id":{"type":"number","description":"唯一标识"},"createdAt":{"format":"date-time","type":"string","description":"创建时间"},"updatedAt":{"format":"date-time","type":"string","description":"更新时间"}},"required":["status","userName","ip","id","createdAt","updatedAt"]},"LoginLogVo":{"type":"object","properties":{"status":{"type":"number","description":"状态。1:成功，0: 失败","enum":[1,0],"default":1},"id":{"type":"number","description":"唯一标识"},"createdAt":{"format":"date-time","type":"string","description":"创建时间"},"updatedAt":{"format":"date-time","type":"string","description":"更新时间"},"userName":{"type":"string","description":"登录账号"},"ip":{"type":"string","description":"登陆地址"},"equipment":{"type":"string","description":"登录设备"},"info":{"type":"string","description":"登录信息"}},"required":["status","id","createdAt","updatedAt","userName","ip"]},"LoginLogListVo":{"type":"object","properties":{"total":{"type":"number","description":"列表总条数"},"pageSize":{"type":"number","description":"分页大小"},"pageNo":{"type":"number","description":"页码"},"list":{"description":"列表数据。","type":"array","items":{"$ref":"#/components/schemas/LoginLogVo"}}},"required":["total","pageSize","pageNo","list"]},"Cpu":{"type":"object","properties":{"model":{"type":"string","description":"model"},"speed":{"type":"number","description":"速度"},"rate":{"type":"number","description":"使用率"}},"required":["model","speed","rate"]},"NetworkInterfaceInfo":{"type":"object","properties":{"address":{"type":"string","description":"地址"},"netmask":{"type":"string","description":"网络掩码"}},"required":["address","netmask"]},"ServerInfoVo":{"type":"object","properties":{"cpus":{"description":"cpu","type":"array","items":{"$ref":"#/components/schemas/Cpu"}},"networks":{"description":"网络","type":"array","items":{"$ref":"#/components/schemas/NetworkInterfaceInfo"}},"arch":{"type":"string","description":"cpu架构"},"kernel":{"type":"string","description":"操作系统内核"},"platform":{"type":"string","description":"操作系统平台"},"uptime":{"type":"number","description":"系统正常运行时间。单位秒"},"hostname":{"type":"string","description":"主机名"},"totalMem":{"type":"number","description":"总内存。单位字节"},"freeMem":{"type":"number","description":"空闲内存。单位字节"}},"required":["cpus","networks","arch","kernel","platform","uptime","hostname","totalMem","freeMem"]}}}}